Revert "Add sub pixel accuracy for rasterization"
