+++ src/Framebuffer.ts
@@ -1,18 +1,16 @@
 
-import { BlenderScene } from './blender/BlenderScene';
 import { ScaleClipBlitter } from './blitter/ScaleClipBlitter';
 import { ControllableCamera } from './camera';
 import { Color } from './core/Color';
 import { CullFace } from './CullFace';
 import { Torus } from './geometrical-objects/Torus';
-import { Matrix3f, Matrix4f, Vector2f, Vector3f, Vector4f } from './math';
+import { Matrix3f, Matrix4f, Vector3f, Vector4f } from './math';
 import { ComputationalGeometryUtils } from './math/Geometry';
 import { Sphere } from './math/Sphere';
 import RandomNumberGenerator from './RandomNumberGenerator';
-import { FlatShadingTriangleRasterizer } from './rasterizer/triangle/FlatShadingTriangleRasterizer';
-import { TexturedTriangleRasterizer } from './rasterizer/triangle/TexturedTriangleRasterizer';
+import { TexturedTriangleRasterizer } from './rasterizer/TexturedTriangleRasterizer';
+import { FlatShadingTriangleRasterizer } from './rasterizer/FlatShadingTriangleRasterizer';
 import { FlatShadingRenderingPipeline } from './rendering-pipelines/FlatShadingRenderingPipeline';
-import { TexturedMesh } from './rendering-pipelines/TexturedMesh';
 import { TexturingRenderingPipeline } from './rendering-pipelines/TexturingRenderingPipeline';
 import { AbstractClipEdge } from './screen-space-clipping/AbstractClipEdge';
 import { BottomClipEdge } from './screen-space-clipping/BottomClipEdge';
@@ -23,15 +21,16 @@ import { TopClipEdge } from './screen-space-clipping/TopClipEdge';
 import { Texture } from './texture/Texture';
 import { TextureCoordinate } from './TextureCoordinate';
 import { Vertex } from "./Vertex";
-import { LineRasterizerDda } from './rasterizer/line/LineRasterizer';
-import { LineRasterizerNoZ } from './rasterizer/line/LineRasterizerNoZ';
-import { BlenderJsonParser } from './blender/BlenderJsonParser';
+import { BlenderScene } from './blender/BlenderScene';
+import { TexturedMesh } from './rendering-pipelines/TexturedMesh';
 
-
+//let bunnyJson = <any>require('./assets/bunny.json');
 // let roomJson = <any>require('./assets/room.json');
 // let hoodlumJson = <any>require('./assets/hoodlum.json');
+//let labJson = <any>require('./assets/lab.json');
 // let labJson2 = <any>require('./assets/lab2.json');
 // let bakedJson = <any>require('./assets/abstract.json');
+//let hlm2018Json = <any>require('./assets/hoodlum2018.json');
 
 export class Framebuffer {
 
@@ -53,6 +52,8 @@ export class Framebuffer {
     private blenderObj: any;
     private blenderObj4: any;
     private blenderObj5: any;
+    private blenderObj6: any;
+    private blenderObj7: any;
     public bob: Texture;
     private sphere: any;
     private plane: any;
@@ -68,8 +69,6 @@ export class Framebuffer {
     public scaleClipBlitter = new ScaleClipBlitter(this);
     public renderingPipeline = new FlatShadingRenderingPipeline(this);
     public texturedRenderingPipeline = new TexturingRenderingPipeline(this);
-    public lineRasterizer = new LineRasterizerDda(this);
-    public lineRasterizerNo = new LineRasterizerNoZ(this);
 
     public setCullFace(face: CullFace): void {
         this.cullMode = face;
@@ -91,9 +90,14 @@ export class Framebuffer {
     }
 
     public precompute(texture: Texture, texture2: Texture): void {
+        //this.bunnyObj = this.createBunny();
         //this.blenderObj4 = this.getBlenderScene(roomJson, false);
         //this.blenderObj5 = this.getBlenderScene(hoodlumJson, false);
+        //this.blenderObj6 = this.getBlenderScene(labJson, false);
+        //this.blenderObj7 = this.getBlenderScene(hlm2018Json, false);
+
         //this.sphere = this.createSphere();
+
         //this.plane = this.createPlane();
         //this.cylinder = this.createCylinder();
         //this.cylinder2 = this.createCylinder2(texture);
@@ -364,7 +368,21 @@ export class Framebuffer {
 
     tmpGlitch = new Uint32Array(320 * 200);
 
+    public drawPolarDistotion(elapsedTime: number, texture: Texture): void {
+        let i = 0;
+        for (let y = 0; y < 200; y++) {
+            for (let x = 0; x < 320; x++) {
+                let xdist = (x - 320 / 2);
+                let ydist = (y - 200 / 2);
+                let dist = Math.sqrt(xdist * xdist + ydist * ydist) * 1.355;
+                let angle = Math.atan2(xdist, ydist) / (Math.PI * 2) * 256;
 
+                let color1 = texture.texture[(dist & 0xff) + (angle & 0xff) * 256];
+
+                this.framebuffer[i++] = color1;
+            }
+        }
+    }
 
     public drawPolarDistotion2(elapsedTime: number, texture: Texture): void {
         let i = 0;
@@ -584,82 +602,7 @@ export class Framebuffer {
                 if (this.wBuffer[index2] > z) {
 
                     let textureIndex = Math.min(xx | 0, texture.width - 1) + Math.min(yy | 0, spritH - 1) * texture.width +
-                        spritH * texture.width * imgNum;
-
-                    let alpha = (texture.texture[textureIndex] >> 24 & 0xff) * alphaScale;
-                    let inverseAlpha = 1 - alpha;
-                    let framebufferPixel = this.framebuffer[index2];
-                    let texturePixel = texture.texture[textureIndex];
-
-                    let r = (framebufferPixel >> 0 & 0xff) * inverseAlpha + (texturePixel >> 0 & 0xff) * alpha;
-                    let g = (framebufferPixel >> 8 & 0xff) * inverseAlpha + (texturePixel >> 8 & 0xff) * alpha;
-                    let b = (framebufferPixel >> 16 & 0xff) * inverseAlpha + (texturePixel >> 16 & 0xff) * alpha;
-
-                    this.framebuffer[index2] = r | (g << 8) | (b << 16) | (255 << 24);
-                }
-                xx += xStep;
-                index2++;
-            }
-            yy += yStep;
-            xx = xTextureStart;
-            index2 += -newWidth + 320;
-        }
-    }
-
-    public drawParticle2Sub(
-        xp: number, yp: number, width: number, height: number, texture: Texture, z: number, alphaBlend: number,
-        imgNum: number = 0, spritH: number): void {
-        let xStep = texture.width / width;
-        let yStep = spritH / height;
-        let xx = 0;
-        let yy = 0;
-
-        let newHeight: number;
-        let newWidth: number
-        let yStart: number;
-        let xStart: number;
-
-        if (yp + height < 0 ||
-            yp > 199 ||
-            xp + width < 0 ||
-            xp > 319) {
-            return;
-        }
-
-        if (yp < 0) {
-            yy = yStep * -yp;
-            newHeight = (height + yp) - Math.max(yp + height - 200, 0);
-            yStart = 0;
-        } else {
-            yStart = yp;
-            newHeight = height - Math.max(yp + height - 200, 0);
-        }
-
-        let xTextureStart: number;
-
-        if (xp < 0) {
-            xTextureStart = xx = xStep * -xp;
-            newWidth = Math.ceil((width + xp) - Math.max(xp + width - 320, 0));
-            xStart = 0;
-        } else {
-            xTextureStart = 0;
-            xStart = xp;
-            newWidth = Math.ceil(width - Math.max(xp + width - 320, 0));
-        }
-
-        const sub: number = Math.ceil(xp) - xp;
-        const suby: number = Math.ceil(yp) - yp;
-        xTextureStart += sub * xStep;
-        yy += suby * yStep;
-
-        const alphaScale = 1 / 255 * alphaBlend;
-        let index2 = Math.ceil(xStart) + Math.ceil(yStart) * 320;
-        for (let y = 0; y < newHeight; y++) {
-            for (let x = 0; x < newWidth; x++) {
-                if (this.wBuffer[index2] > z) {
-
-                    let textureIndex = Math.min(Math.round(xx) | 0, texture.width - 1) + Math.min(Math.round(yy) | 0, spritH - 1) * texture.width +
-                        spritH * texture.width * imgNum;
+                    spritH * texture.width * imgNum;
 
                     let alpha = (texture.texture[textureIndex] >> 24 & 0xff) * alphaScale;
                     let inverseAlpha = 1 - alpha;
@@ -1027,9 +970,9 @@ export class Framebuffer {
                 let framebufferPixel = this.framebuffer[index2];
                 let texturePixel = texture.texture[textureIndex];
 
-                let r = Math.min((framebufferPixel >> 0 & 0xff) + (texturePixel >> 0 & 0xff) * alpha, 255);
-                let g = Math.min((framebufferPixel >> 8 & 0xff) + (texturePixel >> 8 & 0xff) * alpha, 255);
-                let b = Math.min((framebufferPixel >> 16 & 0xff) + (texturePixel >> 16 & 0xff) * alpha, 255);
+                let r = Math.min((framebufferPixel >> 0 & 0xff) + (texturePixel >> 0 & 0xff)*alpha, 255);
+                let g = Math.min((framebufferPixel >> 8 & 0xff) + (texturePixel >> 8 & 0xff)*alpha, 255);
+                let b = Math.min((framebufferPixel >> 16 & 0xff) + (texturePixel >> 16 & 0xff)*alpha, 255);
 
                 this.framebuffer[index2] = r | (g << 8) | (b << 16) | (255 << 24);
                 xx += xStep;
@@ -1336,8 +1279,8 @@ export class Framebuffer {
         }
     }
 
-    public static minWindow: Vector2f = new Vector2f(0, 0);
-    public static maxWindow: Vector2f = new Vector2f(319.0, 199.0);
+    public static minWindow: Vector3f = new Vector3f(0, 0, 0);
+    public static maxWindow: Vector3f = new Vector3f(319, 199, 0);
     /*
         public wireFrameTerrain(elapsedTime: number, heightmap: Texture): void {
 
@@ -1653,6 +1596,81 @@ export class Framebuffer {
         this.drawLensFlare(lensflareScreenSpace, elapsedTime * 0.15, texture, dirt);
     }
 
+    public drawBlenderScene6(elapsedTime: number, texture3: Texture, texture: { tex: Texture, scale: number, alpha: number }[], dirt: Texture): void {
+
+        this.clearDepthBuffer();
+
+        let camera: Matrix4f =
+            Matrix4f.constructTranslationMatrix(0, 0, -34 + (Math.sin(elapsedTime * 0.00007) * 0.5 + 0.5) * 7).multiplyMatrix(
+                Matrix4f.constructXRotationMatrix((Math.sin(elapsedTime * 0.00014) * 0.5 + 0.5) * 0.5 - 0.2).multiplyMatrix(
+                    Matrix4f.constructYRotationMatrix(-elapsedTime * 0.0002).multiplyMatrix(
+
+                        Matrix4f.constructTranslationMatrix(0, 1.9, 0)
+                    )));
+
+
+        let mv: Matrix4f = camera.multiplyMatrix(Matrix4f.constructScaleMatrix(13, 13, 13));
+
+        let scal = Math.sin(elapsedTime * 0.003) * 0.5 + 0.5;
+        for (let j = 0; j < this.blenderObj6.length; j++) {
+            let model = this.blenderObj6[j];
+            this.renderingPipeline.draw(model, mv);
+        }
+
+        mv = camera.multiplyMatrix(
+            Matrix4f.constructTranslationMatrix(0, -5.5, 0).multiplyMatrix(
+                Matrix4f.constructScaleMatrix(413, 413, 413).multiplyMatrix(
+                    Matrix4f.constructXRotationMatrix(Math.PI * 0.5)
+                )
+            ));
+
+        let model = this.blenderObj7[0];
+        this.renderingPipeline.draw(model, mv);
+
+        let points: Array<Vector3f> = new Array<Vector3f>();
+        const num = 10;
+        const num2 = 6;
+
+        for (let i = 0; i < num; i++) {
+
+            for (let j = 0; j < num2; j++) {
+                let y = ((i + elapsedTime * 0.001) % 10) * 2.5 - 12;
+                let scale2 = (1 + 4 * this.interpolate(-10, 10, y)) *
+
+                    ((Math.sin(elapsedTime * 0.0012 + Math.PI * 2 / num * i * 2) * 0.5 + 0.5) * 0.5 + 0.5);
+                let x = scale2 * Math.sin(Math.PI * 2 / num2 * j + elapsedTime * 0.0008);
+
+                let z = scale2 * Math.cos(Math.PI * 2 / num2 * j + elapsedTime * 0.0008);
+
+                points.push(new Vector3f(x, y, z));
+            }
+        }
+
+
+        let modelViewMartrix = camera.multiplyMatrix(Matrix4f.constructTranslationMatrix(0, -0.0, 0));
+
+        let points2: Array<Vector3f> = new Array<Vector3f>(points.length);
+        points.forEach(element => {
+
+
+            let transformed = this.project(modelViewMartrix.multiply(element));
+
+            points2.push(transformed);
+        });
+
+        points2.sort(function (a, b) {
+            return a.z - b.z;
+        });
+
+        points2.forEach(element => {
+            let size = -(4.3 * 192 / (element.z));
+            this.drawSoftParticle(
+                Math.round(element.x - size / 2),
+                Math.round(element.y - size / 2),
+                Math.round(size), Math.round(size), texture3, 1 / element.z, 0.7);
+        });
+    }
+
     public drawPlaneDeformation(elapsedTime: number, texture: Texture): void {
         // optimize
         // power of two modulo with &
@@ -1687,6 +1705,48 @@ export class Framebuffer {
                 this.framebuffer[framebufferIndex++] = color;
             }
         }
+    }
+
+    /**
+     * this rountine is pretty slow:
+     * - optimize scaled blittinh
+     * - optimize geometry stage by reusing arrays
+     * - dont us forEach!
+     */
+    public scene7(elapsedTime: number, texture: Texture): void {
+        let points: Array<Vector3f> = new Array<Vector3f>();
+
+        for (let i = 0; i < 120; i++) {
+            points.push(new Vector3f(Math.sin(i * 0.25) * 8, i * 0.3 - 18, Math.cos(i * 0.25) * 8));
+        }
+
+        points.push(new Vector3f(0, 0, 5));
+
+        let rotMat = Matrix3f.constructYRotationMatrix(elapsedTime * 0.0005);
+        rotMat = rotMat.multiplyMatrix(Matrix3f.constructXRotationMatrix(elapsedTime * 0.0002));
+
+        let points2: Array<Vector3f> = new Array<Vector3f>(points.length);
+        points.forEach(element => {
+            let alpha = -elapsedTime * 0.0013;
+
+            let transformed = rotMat.multiply(element);
+
+            let x = transformed.x;
+            let y = transformed.y;
+            let z = transformed.z - 10;
+            let xx = 320 / 2 + (x / (z * 0.0058));
+            let yy = 200 / 2 + (y / (z * 0.0058));
+            points2.push(new Vector3f(xx, yy, z));
+        });
+
+        points2.sort(function (a, b) {
+            return a.z - b.z;
+        });
+
+        points2.forEach(element => {
+            let size = -(1.9 / (element.z * 0.0058)) | 0;
+            this.drawSoftParticle((element.x - size / 2) | 0, (element.y - size / 2) | 0, size, size, texture, 1 / element.z, 1.0);
+        });
     }
 
     public shadingSphereClip(elapsedTime: number): void {
@@ -2955,13 +3015,125 @@ export class Framebuffer {
             }
         }
     */
+    /*
+        public createBunny(): any {
+            let points: Array<Vector3f> = new Array<Vector3f>();
 
+            bunnyJson.vertices.forEach(x => {
+                points.push(new Vector3f(x.x, x.y, x.z));
+            });
+
+            let normals: Array<Vector3f> = new Array<Vector3f>();
+
+            bunnyJson.normals.forEach(x => {
+                normals.push(new Vector3f(x.x, x.y, x.z).normalize());
+            });
+
+            let index: Array<number> = bunnyJson.faces;
+
+            let points2: Array<Vector3f> = new Array<Vector3f>();
+            let normals2: Array<Vector3f> = new Array<Vector3f>();
+
+            for (let i = 0; i < points.length; i++) {
+                points2.push(new Vector3f(0, 0, 0));
+            }
+
+            for (let i = 0; i < normals.length; i++) {
+                normals2.push(new Vector3f(0, 0, 0));
+            }
+
+            let object = {
+                index: index,
+                points: points,
+                normals: normals,
+                points2: points2,
+                normals2: normals2
+            };
+
+            return object;
+        }
+    */
+    /*
+    public reflectionBunny(elapsedTime: number): void {
+        this.clearDepthBuffer();
+
+        let obj = this.bunnyObj;
+
+        let scale = 64.1;
+        let modelViewMartrix = Matrix4f.constructScaleMatrix(scale, scale, scale).multiplyMatrix(Matrix4f.constructYRotationMatrix(elapsedTime * 0.30));
+        modelViewMartrix = modelViewMartrix.multiplyMatrix(Matrix4f.constructXRotationMatrix(elapsedTime * 0.3));
+        modelViewMartrix = Matrix4f.constructTranslationMatrix(0, 0, -8).multiplyMatrix(modelViewMartrix);
+
+        let normalMatrix = modelViewMartrix.computeNormalMatrix();
+
+        for (let n = 0; n < obj.normals.length; n++) {
+            normalMatrix.multiplyArr(obj.normals[n], obj.normals2[n]);
+        }
+
+        for (let p = 0; p < obj.points.length; p++) {
+            let transformed = modelViewMartrix.multiply(obj.points[p]);
+
+            let x = transformed.x;
+            let y = transformed.y;
+            let z = transformed.z; // TODO: use translation matrix!
+
+            let xx = (320 * 0.5) + (x / (-z * 0.0078));
+            let yy = (200 * 0.5) - (y / (-z * 0.0078));
+
+            obj.points2[p].x = Math.round(xx);
+            obj.points2[p].y = Math.round(yy);
+            obj.points2[p].z = z;
+        }
+
+        let vertex1 = new Vertex();
+        vertex1.textureCoordinate = new TextureCoordinate();
+        let vertex2 = new Vertex();
+        vertex2.textureCoordinate = new TextureCoordinate();
+        let vertex3 = new Vertex();
+        vertex3.textureCoordinate = new TextureCoordinate();
+        let vertexArray = new Array<Vertex>(vertex1, vertex2, vertex3);
+
+        for (let i = 0; i < obj.index.length; i += 6) {
+            let v1 = obj.points2[obj.index[i]];
+            let v2 = obj.points2[obj.index[i + 1]];
+            let v3 = obj.points2[obj.index[i + 2]];
+
+            if (this.isTriangleCCW(v1, v2, v3)) {
+                vertexArray[0].position = v1;
+                this.fakeSphere(obj.normals2[obj.index[i + 3]], vertex1);
+
+                vertexArray[1].position = v2;
+                this.fakeSphere(obj.normals2[obj.index[i + 4]], vertex2);
+
+                vertexArray[2].position = v3;
+                this.fakeSphere(obj.normals2[obj.index[i + 5]], vertex3);
+
+                if (v1.x < Framebuffer.minWindow.x ||
+                    v2.x < Framebuffer.minWindow.x ||
+                    v3.x < Framebuffer.minWindow.x ||
+                    v1.x > Framebuffer.maxWindow.x ||
+                    v2.x > Framebuffer.maxWindow.x ||
+                    v3.x > Framebuffer.maxWindow.x ||
+                    v1.y < Framebuffer.minWindow.y ||
+                    v2.y < Framebuffer.minWindow.y ||
+                    v3.y < Framebuffer.minWindow.y ||
+                    v1.y > Framebuffer.maxWindow.y ||
+                    v2.y > Framebuffer.maxWindow.y ||
+                    v3.y > Framebuffer.maxWindow.y) {
+                    this.clipConvexPolygon2(vertexArray, 0);
+                } else {
+                    this.texturedTriangleRasterizer.drawTriangleDDA2(vertexArray[0], vertexArray[1], vertexArray[2], 0);
+                }
+            }
+        }
+    }
+*/
     public fakeSphere(normal: Vector4f, vertex: Vertex): void {
         // https://www.mvps.org/directx/articles/spheremap.htm
-        vertex.textureCoordinate.u = 0.5 + normal.x * 0.5;
-        vertex.textureCoordinate.v = 0.5 - normal.y * 0.5;
-        //vertex.textureCoordinate.u = 0.5 + Math.asin(normal.x) / Math.PI;
-        //vertex.textureCoordinate.v = 0.5 - Math.asin(normal.y) / Math.PI;
+        //vertex.textureCoordinate.u = 0.5 + normal.x * 0.5;
+        //vertex.textureCoordinate.v = 0.5 - normal.y * 0.5;
+        vertex.textureCoordinate.u = 0.5 + Math.asin(normal.x) / Math.PI;
+        vertex.textureCoordinate.v = 0.5 - Math.asin(normal.y) / Math.PI;
     }
 
     public fakeSphere2(normal: Vector3f, tex: TextureCoordinate): void {
@@ -2976,9 +3148,10 @@ export class Framebuffer {
         new TopClipEdge()
     );
 
-    private lensFlareVisible: boolean = false;
-    private lensFlareStart = 0;
-    private lensFlareEnd = 0;
+
+    lensFlareVisible: boolean = false;
+    lensFlareStart = 0;
+    lensFlareEnd = 0;
 
     public drawLensFlare(screenPos: Vector3f, elapsedTime: number, texture: { tex: Texture, scale: number, alpha: number }[], dirt: Texture): void {
         let pos = screenPos;
@@ -3014,12 +3187,127 @@ export class Framebuffer {
         this.drawTextureRectAdd(0, 0, 0, 0, 320, 200, dirt, 0.03 + 0.15 * scale);
     }
 
+    /**
+     * digital differential analyser line drawing algorithm
+     * using fixed point math.
+     * renders approx 1400 lines per millisecond on my machine
+     */
     public drawLineDDA(start: Vector3f, end: Vector3f, color: number): void {
-        this.lineRasterizer.drawLineDDA(start, end, color);
+        let xDistance: number = end.x - start.x;
+        let yDistance: number = end.y - start.y;
+
+        let dx: number, dy: number, length: number;
+
+        if (Math.abs(xDistance) > Math.abs(yDistance)) {
+            dx = Math.sign(xDistance);
+            dy = yDistance / Math.abs(xDistance);
+            length = Math.abs(xDistance);
+        } else {
+            dx = xDistance / Math.abs(yDistance);
+            dy = Math.sign(yDistance);
+            length = Math.abs(yDistance);
+        }
+
+        let xPosition: number = start.x;
+        let yPosition: number = start.y;
+
+        // w=1/z interpolation for z-buffer
+        let wStart = 1 / (start.z);
+        let wDelta = (1 / end.z - 1 / start.z) / length;
+
+        for (let i = 0; i <= length; i++) {
+            if (wStart < this.wBuffer[Math.round(xPosition) + Math.round(yPosition) * 320]) {
+                this.wBuffer[Math.round(xPosition) + Math.round(yPosition) * 320] = wStart;
+                this.drawPixel(Math.round(xPosition), Math.round(yPosition), color);
+            }
+            xPosition += dx;
+            yPosition += dy;
+            wStart += wDelta;
+        }
     }
 
     public drawLineDDANoZ(start: Vector3f, end: Vector3f, color: number): void {
-        this.lineRasterizerNo.drawLineDDANoZ(start, end, color);
+        let xDistance: number = end.x - start.x;
+        let yDistance: number = end.y - start.y;
+
+        let dx: number, dy: number, length: number;
+
+        if (Math.abs(xDistance) > Math.abs(yDistance)) {
+            dx = Math.sign(xDistance);
+            dy = yDistance / Math.abs(xDistance);
+            length = Math.abs(xDistance);
+        } else {
+            dx = xDistance / Math.abs(yDistance);
+            dy = Math.sign(yDistance);
+            length = Math.abs(yDistance);
+        }
+
+        let xPosition: number = start.x;
+        let yPosition: number = start.y;
+
+        for (let i = 0; i <= length; i++) {
+
+            this.drawPixel(Math.round(xPosition), Math.round(yPosition), color);
+
+            xPosition += dx;
+            yPosition += dy;
+
+        }
+    }
+
+    drawVoxelLandscape3(texture: Texture, time: number) {
+        this.clearColorBuffer(255 << 24);
+
+        const MIN_DIST = 10;
+        const MAX_DIST = 100;
+
+        let camX = time * 0.006;
+        let camY = 0;
+
+        const focus = 28.7;
+        const center = 220;
+        const eye = 120;
+
+        for (let x = 0; x < 320; x++) {
+            let dirX = Math.cos(time * 0.0005 + x * 0.005) * 0.4;
+            let dirY = Math.sin(time * 0.0005 + x * 0.005) * 0.4;
+
+            dirX = Math.cos(time * 0.0001 + Math.PI * 2 / 320 * x) * 0.4;
+            dirY = Math.sin(time * 0.0001 + Math.PI * 2 / 320 * x) * 0.4;
+
+            let highestPoint = 0;
+
+            let rayX = camX + dirX * MIN_DIST;
+            let rayY = camY + dirY * MIN_DIST;
+
+            for (let dist = MIN_DIST; dist < MAX_DIST; dist++) {
+
+                let height = //this.getBilinearFilteredPixel(texture, rayX, rayY)*0.6;
+                    (texture.texture[(rayX & 0xff) + (rayY & 0xff) * 256] & 0xff) * 0.6;
+                let projHeight = Math.round((height - eye) * focus / dist + center) - 50;
+                let color = (Math.round(height) * 200 / 255 + 55) * Math.min(1.0, (1 - (dist - MIN_DIST) / (MAX_DIST - MIN_DIST)));
+                let packedRGB = 255 << 24 | (color * 0.7) << 16 | (color) << 8 | (color * 0.8);
+
+                if (projHeight > highestPoint) {
+                    let index = x + (199 - highestPoint) * 320;
+                    let max = Math.min(projHeight, 200);
+
+                    for (let i = highestPoint; i < max; i++) {
+                        this.framebuffer[index] = packedRGB;
+                        index -= 320;
+                    }
+
+                    if (max == 200) {
+                        break;
+                    }
+
+                    highestPoint = projHeight;
+                }
+
+                rayX += dirX;
+                rayY += dirY;
+            }
+        }
     }
 
 }
