--- src/examples/lab-scene/Application.ts
@@ -1,14 +1,0 @@
-import { Canvas } from '../../Canvas';
-import { MovingTorusScene } from './MovingTorusScene';
-
-class Application {
-
-    public static main(): void {
-        const canvas: Canvas = new Canvas(320, 200, new MovingTorusScene());
-        canvas.appendTo(document.getElementById('aisa'));
-        canvas.init();
-    }
-
-}
-
-Application.main();
