--- src/examples/lab-scene/MovingTorusScene.ts
@@ -1,158 +1,0 @@
-import { CullFace } from '../../CullFace';
-import { Framebuffer } from '../../Framebuffer';
-import { Torus } from '../../geometrical-objects/Torus';
-import { Matrix4f, Vector3f } from '../../math/index';
-import { AbstractScene } from '../../scenes/AbstractScene';
-import { Texture, TextureUtils } from '../../texture/index';
-import RandomNumberGenerator from '../../RandomNumberGenerator';
-import { FlatShadingRenderingPipeline } from '../../rendering-pipelines/FlatShadingRenderingPipeline';
-import { BlenderJsonParser } from '../../blender/BlenderJsonParser';
-import { FlatshadedMesh } from '../../geometrical-objects/FlatshadedMesh';
-
-const labJson = <any> require('./../../assets/lab.json');
-
-const hlm2018Json = <any>require('./../../assets/hoodlum2018.json');
-
-export class MovingTorusScene extends AbstractScene {
-
-    private noise: Texture;
-    private accumulationBuffer: Uint32Array = new Uint32Array(320 * 200);
-    private particleTexture2: Texture;
-    private blenderObj6: Array<FlatshadedMesh>;
-    private blenderObj7: Array<FlatshadedMesh>;
-
-    private renderingPipeline: FlatShadingRenderingPipeline;
-
-    public init(framebuffer: Framebuffer): Promise<any> {
-        this.renderingPipeline = new FlatShadingRenderingPipeline(framebuffer);
-        this.blenderObj6 = BlenderJsonParser.parse(labJson, false);
-        this.blenderObj7 = BlenderJsonParser.parse(hlm2018Json, false);
-        return Promise.all([
-            TextureUtils.generateProceduralNoise().then(
-                (texture: Texture) => this.noise = texture
-            ),
-            this.createProceduralTexture3().then(texture => this.particleTexture2 = texture)
-        ]);
-    }
-
-    public createProceduralTexture3(): Promise<Texture> {
-        return new Promise((resolve) => {
-            const texture = new Texture();
-            texture.texture = new Uint32Array(256 * 256);
-
-            let rng = new RandomNumberGenerator();
-            rng.setSeed(100);
-
-            for (let y = 0; y < 256; y++) {
-                for (let x = 0; x < 256; x++) {
-                    let dx = 127 - x
-                    let dy = 127 - y
-                    let r = Math.sqrt(dx * dx + dy * dy) / 127;
-                    let c = 1 - r;
-                    c = c * c * c;
-                    if (r > 1) c = 0;
-                    c = Math.min(1, c * 2.9);
-
-                    texture.texture[x + y * 256] = 235 | 255 << 8 | 235 << 16 | (c * 255) << 24;
-                }
-            }
-
-            texture.width = 256;
-            texture.height = 256;
-            resolve(texture);
-        });
-    }
-
-    public render(framebuffer: Framebuffer): void {
-        const time: number = Date.now();
-
-        framebuffer.setCullFace(CullFace.BACK);
-
-
-        this.drawBlenderScene6(framebuffer, time);
-
-        const texture3: Texture = new Texture(this.accumulationBuffer, 320, 200);
-        framebuffer.drawTexture(0, 0, texture3, 0.85);
-        framebuffer.fastFramebufferCopy(this.accumulationBuffer, framebuffer.framebuffer);
-
-        framebuffer.noise(time, this.noise);
-
-    }
-
-    public drawBlenderScene6(framebuffer: Framebuffer, elapsedTime: number): void {
-
-        framebuffer.clearDepthBuffer();
-
-        let camera: Matrix4f =
-            Matrix4f.constructTranslationMatrix(0, 0, -34 + (Math.sin(elapsedTime * 0.00007) * 0.5 + 0.5) * 7).multiplyMatrix(
-                Matrix4f.constructXRotationMatrix((Math.sin(elapsedTime * 0.00014) * 0.5 + 0.5) * 0.5 - 0.2).multiplyMatrix(
-                    Matrix4f.constructYRotationMatrix(-elapsedTime * 0.0002).multiplyMatrix(
-
-                        Matrix4f.constructTranslationMatrix(0, 1.9, 0)
-                    )));
-
-
-        let mv: Matrix4f = camera.multiplyMatrix(Matrix4f.constructScaleMatrix(13, 13, 13));
-
-        let scal = Math.sin(elapsedTime * 0.003) * 0.5 + 0.5;
-        for (let j = 0; j < this.blenderObj6.length; j++) {
-            let model = this.blenderObj6[j];
-            this.renderingPipeline.draw(model, mv);
-        }
-
-        mv = camera.multiplyMatrix(
-            Matrix4f.constructTranslationMatrix(0, -5.5, 0).multiplyMatrix(
-                Matrix4f.constructScaleMatrix(413, 413, 413).multiplyMatrix(
-                    Matrix4f.constructXRotationMatrix(Math.PI * 0.5)
-                )
-            ));
-
-        let model = this.blenderObj7[0];
-        this.renderingPipeline.draw(model, mv);
-
-        let points: Array<Vector3f> = new Array<Vector3f>();
-        const num = 10;
-        const num2 = 6;
-
-        for (let i = 0; i < num; i++) {
-
-            for (let j = 0; j < num2; j++) {
-                let y = ((i + elapsedTime * 0.001) % 10) * 2.5 - 12;
-                let scale2 = (1 + 4 * framebuffer.interpolate(-10, 10, y)) *
-
-                    ((Math.sin(elapsedTime * 0.0012 + Math.PI * 2 / num * i * 2) * 0.5 + 0.5) * 0.5 + 0.5);
-                let x = scale2 * Math.sin(Math.PI * 2 / num2 * j + elapsedTime * 0.0008);
-
-                let z = scale2 * Math.cos(Math.PI * 2 / num2 * j + elapsedTime * 0.0008);
-
-                points.push(new Vector3f(x, y, z));
-            }
-        }
-
-
-        let modelViewMartrix = camera.multiplyMatrix(Matrix4f.constructTranslationMatrix(0, -0.0, 0));
-
-        let points2: Array<Vector3f> = new Array<Vector3f>(points.length);
-        points.forEach(element => {
-
-
-            let transformed = framebuffer.project(modelViewMartrix.multiply(element));
-
-            points2.push(transformed);
-        });
-
-        points2.sort(function (a, b) {
-            return a.z - b.z;
-        });
-
-        points2.forEach(element => {
-            let size = -(4.3 * 192 / (element.z));
-            framebuffer.drawSoftParticle(
-                Math.round(element.x - size / 2),
-                Math.round(element.y - size / 2),
-                Math.round(size), Math.round(size), this.particleTexture2, 1 / element.z, 0.7);
-        });
-    }
-
-
-}
