+++ src/rasterizer/GouraudShadingTriangleRasterizer.ts
@@ -1,10 +1,16 @@
-import { Framebuffer } from '../../Framebuffer';
-import { Vertex } from '../../Vertex';
+import { Framebuffer } from '../Framebuffer';
+import { Vertex } from '../Vertex';
 import { AbstractTriangleRasterizer } from './AbstractTriangleRasterizer';
+import { ColorInterpolator } from './ColorInterpolator';
 
-export class FlatShadingTriangleRasterizer extends AbstractTriangleRasterizer {
+export class GouraudShadingTriangleRasterizer extends AbstractTriangleRasterizer {
 
     private temp: Vertex = null;
+
+    private colorInterpolator1: ColorInterpolator = new ColorInterpolator();
+    private colorInterpolator2: ColorInterpolator = new ColorInterpolator();
+    private colorInterpolator3: ColorInterpolator = new ColorInterpolator();
+    private rowColorInterpolator: ColorInterpolator = new ColorInterpolator();
 
     constructor(private framebuffer: Framebuffer) {
         super();
@@ -61,12 +67,13 @@ export class FlatShadingTriangleRasterizer extends AbstractTriangleRasterizer {
     }
 
     private fillBottomFlatTriangle(v1: Vertex, v2: Vertex, v3: Vertex): void {
-        const color: number = v1.color.toPackedFormat();
-
         const yDistance: number = v3.projection.y - v1.projection.y;
 
         const slope1: number = (v2.projection.x - v1.projection.x) / yDistance;
         const slope2: number = (v3.projection.x - v1.projection.x) / yDistance;
+
+        this.colorInterpolator1.setup(v1.color, v2.color, yDistance);
+        this.colorInterpolator2.setup(v1.color, v3.color, yDistance);
 
         const zslope1: number = (1 / v2.projection.z - 1 / v1.projection.z) / yDistance;
         const zslope2: number = (1 / v3.projection.z - 1 / v1.projection.z) / yDistance;
@@ -80,33 +87,49 @@ export class FlatShadingTriangleRasterizer extends AbstractTriangleRasterizer {
 
         for (let i = 0; i < yDistance; i++) {
             let length = Math.round(xPosition2) - Math.round(xPosition);
+            this.rowColorInterpolator.setup(
+                this.colorInterpolator1.startColor, this.colorInterpolator2.startColor, length);
             let framebufferIndex = Math.round(yPosition) * 320 + Math.round(xPosition);
             let spanzStep = (curz2 - curz1) / length;
             let wStart = curz1;
             for (let j = 0; j < length; j++) {
                 if (wStart < this.framebuffer.wBuffer[framebufferIndex]) {
                     this.framebuffer.wBuffer[framebufferIndex] = wStart;
-                    this.framebuffer.framebuffer[framebufferIndex] = color;
+                    this.framebuffer.framebuffer[framebufferIndex] =
+                        this.rowColorInterpolator.startColor.toPackedFormat();
                 }
                 framebufferIndex++;
                 wStart += spanzStep;
+                this.rowColorInterpolator.startColor.r += this.rowColorInterpolator.colorSlope.r;
+                this.rowColorInterpolator.startColor.g += this.rowColorInterpolator.colorSlope.g;
+                this.rowColorInterpolator.startColor.b += this.rowColorInterpolator.colorSlope.b;
             }
 
             xPosition += slope1;
             xPosition2 += slope2;
             yPosition++;
 
+            this.colorInterpolator1.startColor.r += this.colorInterpolator1.colorSlope.r;
+            this.colorInterpolator1.startColor.g += this.colorInterpolator1.colorSlope.g;
+            this.colorInterpolator1.startColor.b += this.colorInterpolator1.colorSlope.b;
+
+            this.colorInterpolator2.startColor.r += this.colorInterpolator2.colorSlope.r;
+            this.colorInterpolator2.startColor.g += this.colorInterpolator2.colorSlope.g;
+            this.colorInterpolator2.startColor.b += this.colorInterpolator2.colorSlope.b;
+
             curz1 += zslope1;
             curz2 += zslope2;
         }
     }
 
-    fillTopFlatTriangle(v1: Vertex, v2: Vertex, v3: Vertex): void {
-        const color: number = v1.color.toPackedFormat();
-
+    private fillTopFlatTriangle(v1: Vertex, v2: Vertex, v3: Vertex): void {
         let yDistance = v3.projection.y - v1.projection.y;
         let slope1 = (v3.projection.x - v1.projection.x) / yDistance;
         let slope2 = (v3.projection.x - v2.projection.x) / yDistance;
+
+
+        this.colorInterpolator1.setup(v1.color, v3.color, yDistance);
+        this.colorInterpolator2.setup(v2.color, v3.color, yDistance);
 
         let zslope1 = (1 / v3.projection.z - 1 / v1.projection.z) / yDistance;
         let zslope2 = (1 / v3.projection.z - 1 / v2.projection.z) / yDistance;
@@ -120,30 +143,45 @@ export class FlatShadingTriangleRasterizer extends AbstractTriangleRasterizer {
 
         for (let i = 0; i < yDistance; i++) {
             let length = Math.round(xPosition2) - Math.round(xPosition);
+            this.rowColorInterpolator.setup(
+                this.colorInterpolator1.startColor, this.colorInterpolator2.startColor, length);
             let framebufferIndex = Math.round(yPosition) * 320 + Math.round(xPosition);
             for (let j = 0; j < length; j++) {
                 let wStart = (curz2 - curz1) / (length) * j + curz1;
                 if (wStart < this.framebuffer.wBuffer[framebufferIndex]) {
                     this.framebuffer.wBuffer[framebufferIndex] = wStart;
-                    this.framebuffer.framebuffer[framebufferIndex] = color;
+                    this.framebuffer.framebuffer[framebufferIndex] = this.rowColorInterpolator.startColor.toPackedFormat();
                 }
                 framebufferIndex++;
+                this.rowColorInterpolator.startColor.r += this.rowColorInterpolator.colorSlope.r;
+                this.rowColorInterpolator.startColor.g += this.rowColorInterpolator.colorSlope.g;
+                this.rowColorInterpolator.startColor.b += this.rowColorInterpolator.colorSlope.b;
             }
 
             xPosition += slope1;
             xPosition2 += slope2;
             yPosition++;
 
+            this.colorInterpolator1.startColor.r += this.colorInterpolator1.colorSlope.r;
+            this.colorInterpolator1.startColor.g += this.colorInterpolator1.colorSlope.g;
+            this.colorInterpolator1.startColor.b += this.colorInterpolator1.colorSlope.b;
+
+            this.colorInterpolator2.startColor.r += this.colorInterpolator2.colorSlope.r;
+            this.colorInterpolator2.startColor.g += this.colorInterpolator2.colorSlope.g;
+            this.colorInterpolator2.startColor.b += this.colorInterpolator2.colorSlope.b;
+
             curz1 += zslope1;
             curz2 += zslope2;
         }
     }
 
-    fillLongRightTriangle(v1: Vertex, v2: Vertex, v3: Vertex): void {
-        const color: number = v1.color.toPackedFormat();
-
+    private fillLongRightTriangle(v1: Vertex, v2: Vertex, v3: Vertex): void {
         let yDistanceLeft = v2.projection.y - v1.projection.y;
         let yDistanceRight = v3.projection.y - v1.projection.y;
+
+        this.colorInterpolator1.setup(v1.color, v2.color, yDistanceLeft);
+        this.colorInterpolator2.setup(v1.color, v3.color, yDistanceRight);
+
 
         let slope1 = (v2.projection.x - v1.projection.x) / yDistanceLeft;
         let slope2 = (v3.projection.x - v1.projection.x) / yDistanceRight;
@@ -160,27 +198,41 @@ export class FlatShadingTriangleRasterizer extends AbstractTriangleRasterizer {
 
         for (let i = 0; i < yDistanceLeft; i++) {
             let length = Math.round(xPosition2) - Math.round(xPosition);
+            this.rowColorInterpolator.setup(
+                this.colorInterpolator1.startColor, this.colorInterpolator2.startColor, length);
             let framebufferIndex = Math.round(yPosition) * 320 + Math.round(xPosition);
             let spanzStep = (curz2 - curz1) / length;
             let wStart = curz1;
             for (let j = 0; j < length; j++) {
                 if (wStart < this.framebuffer.wBuffer[framebufferIndex]) {
                     this.framebuffer.wBuffer[framebufferIndex] = wStart;
-                    this.framebuffer.framebuffer[framebufferIndex] = color;
+                    this.framebuffer.framebuffer[framebufferIndex] = this.rowColorInterpolator.startColor.toPackedFormat();
                 }
                 framebufferIndex++;
                 wStart += spanzStep;
+                this.rowColorInterpolator.startColor.r += this.rowColorInterpolator.colorSlope.r;
+                this.rowColorInterpolator.startColor.g += this.rowColorInterpolator.colorSlope.g;
+                this.rowColorInterpolator.startColor.b += this.rowColorInterpolator.colorSlope.b;
             }
 
             xPosition += slope1;
             xPosition2 += slope2;
             yPosition++;
 
+            this.colorInterpolator1.startColor.r += this.colorInterpolator1.colorSlope.r;
+            this.colorInterpolator1.startColor.g += this.colorInterpolator1.colorSlope.g;
+            this.colorInterpolator1.startColor.b += this.colorInterpolator1.colorSlope.b;
+
+            this.colorInterpolator2.startColor.r += this.colorInterpolator2.colorSlope.r;
+            this.colorInterpolator2.startColor.g += this.colorInterpolator2.colorSlope.g;
+            this.colorInterpolator2.startColor.b += this.colorInterpolator2.colorSlope.b;
+
             curz1 += zslope1;
             curz2 += zslope2;
         }
 
         yDistanceLeft = v3.projection.y - v2.projection.y;
+        this.colorInterpolator3.setup(v2.color, v3.color, yDistanceLeft);
         slope1 = (v3.projection.x - v2.projection.x) / yDistanceLeft;
         zslope1 = (1 / v3.projection.z - 1 / v2.projection.z) / yDistanceLeft;
 
@@ -189,21 +241,34 @@ export class FlatShadingTriangleRasterizer extends AbstractTriangleRasterizer {
 
         for (let i = 0; i < yDistanceLeft; i++) {
             let length = Math.round(xPosition2) - Math.round(xPosition);
-            let framebufferIndex = Math.round(yPosition) * 320 + Math.round(xPosition)
+            this.rowColorInterpolator.setup(
+                this.colorInterpolator3.startColor, this.colorInterpolator2.startColor, length);
+            let framebufferIndex = Math.round(yPosition) * 320 + Math.round(xPosition);
             let spanzStep = (curz2 - curz1) / length;
             let wStart = curz1;
             for (let j = 0; j < length; j++) {
                 if (wStart < this.framebuffer.wBuffer[framebufferIndex]) {
                     this.framebuffer.wBuffer[framebufferIndex] = wStart;
-                    this.framebuffer.framebuffer[framebufferIndex] = color;
+                    this.framebuffer.framebuffer[framebufferIndex] = this.rowColorInterpolator.startColor.toPackedFormat();
                 }
                 framebufferIndex++;
                 wStart += spanzStep;
+                this.rowColorInterpolator.startColor.r += this.rowColorInterpolator.colorSlope.r;
+                this.rowColorInterpolator.startColor.g += this.rowColorInterpolator.colorSlope.g;
+                this.rowColorInterpolator.startColor.b += this.rowColorInterpolator.colorSlope.b;
             }
 
             xPosition += slope1;
             xPosition2 += slope2;
             yPosition++;
+
+            this.colorInterpolator2.startColor.r += this.colorInterpolator2.colorSlope.r;
+            this.colorInterpolator2.startColor.g += this.colorInterpolator2.colorSlope.g;
+            this.colorInterpolator2.startColor.b += this.colorInterpolator2.colorSlope.b;
+
+            this.colorInterpolator3.startColor.r += this.colorInterpolator3.colorSlope.r;
+            this.colorInterpolator3.startColor.g += this.colorInterpolator3.colorSlope.g;
+            this.colorInterpolator3.startColor.b += this.colorInterpolator3.colorSlope.b;
 
             curz1 += zslope1;
             curz2 += zslope2;
@@ -212,10 +277,11 @@ export class FlatShadingTriangleRasterizer extends AbstractTriangleRasterizer {
 
 
     fillLongLeftTriangle(v1: Vertex, v2: Vertex, v3: Vertex): void {
-        const color: number = v1.color.toPackedFormat();
-
         let yDistanceRight = v2.projection.y - v1.projection.y;
         let yDistanceLeft = v3.projection.y - v1.projection.y;
+
+        this.colorInterpolator2.setup(v1.color, v2.color, yDistanceRight);
+        this.colorInterpolator1.setup(v1.color, v3.color, yDistanceLeft);
 
         let slope2 = (v2.projection.x - v1.projection.x) / yDistanceRight;
         let slope1 = (v3.projection.x - v1.projection.x) / yDistanceLeft;
@@ -232,27 +298,40 @@ export class FlatShadingTriangleRasterizer extends AbstractTriangleRasterizer {
 
         for (let i = 0; i < yDistanceRight; i++) {
             let length = Math.round(xPosition2) - Math.round(xPosition);
+            this.rowColorInterpolator.setup(
+                this.colorInterpolator1.startColor, this.colorInterpolator2.startColor, length);
             let framebufferIndex = Math.round(yPosition) * 320 + Math.round(xPosition);
             let spanzStep = (curz2 - curz1) / length;
             let wStart = curz1;
             for (let j = 0; j < length; j++) {
                 if (wStart < this.framebuffer.wBuffer[framebufferIndex]) {
                     this.framebuffer.wBuffer[framebufferIndex] = wStart;
-                    this.framebuffer.framebuffer[framebufferIndex] = color;
+                    this.framebuffer.framebuffer[framebufferIndex] = this.rowColorInterpolator.startColor.toPackedFormat();
                 }
                 framebufferIndex++;
                 wStart += spanzStep;
+                this.rowColorInterpolator.startColor.r += this.rowColorInterpolator.colorSlope.r;
+                this.rowColorInterpolator.startColor.g += this.rowColorInterpolator.colorSlope.g;
+                this.rowColorInterpolator.startColor.b += this.rowColorInterpolator.colorSlope.b;
             }
 
             xPosition += slope1;
             xPosition2 += slope2;
             yPosition++;
+            this.colorInterpolator1.startColor.r += this.colorInterpolator1.colorSlope.r;
+            this.colorInterpolator1.startColor.g += this.colorInterpolator1.colorSlope.g;
+            this.colorInterpolator1.startColor.b += this.colorInterpolator1.colorSlope.b;
+
+            this.colorInterpolator2.startColor.r += this.colorInterpolator2.colorSlope.r;
+            this.colorInterpolator2.startColor.g += this.colorInterpolator2.colorSlope.g;
+            this.colorInterpolator2.startColor.b += this.colorInterpolator2.colorSlope.b;
 
             curz1 += zslope1;
             curz2 += zslope2;
         }
 
         yDistanceRight = v3.projection.y - v2.projection.y;
+        this.colorInterpolator3.setup(v2.color, v3.color, yDistanceRight);
         slope2 = (v3.projection.x - v2.projection.x) / yDistanceRight;
         zslope2 = (1 / v3.projection.z - 1 / v2.projection.z) / yDistanceRight;
 
@@ -262,21 +341,34 @@ export class FlatShadingTriangleRasterizer extends AbstractTriangleRasterizer {
 
         for (let i = 0; i < yDistanceRight; i++) {
             let length = Math.round(xPosition2) - Math.round(xPosition);
+            this.rowColorInterpolator.setup(
+                this.colorInterpolator1.startColor, this.colorInterpolator3.startColor, length);
             let framebufferIndex = Math.round(yPosition) * 320 + Math.round(xPosition)
             let spanzStep = (curz2 - curz1) / length;
             let wStart = curz1;
             for (let j = 0; j < length; j++) {
                 if (wStart < this.framebuffer.wBuffer[framebufferIndex]) {
                     this.framebuffer.wBuffer[framebufferIndex] = wStart;
-                    this.framebuffer.framebuffer[framebufferIndex] = color;
+                    this.framebuffer.framebuffer[framebufferIndex] = this.rowColorInterpolator.startColor.toPackedFormat();
                 }
                 framebufferIndex++;
                 wStart += spanzStep;
+
+                this.rowColorInterpolator.startColor.r += this.rowColorInterpolator.colorSlope.r;
+                this.rowColorInterpolator.startColor.g += this.rowColorInterpolator.colorSlope.g;
+                this.rowColorInterpolator.startColor.b += this.rowColorInterpolator.colorSlope.b;
             }
 
             xPosition += slope1;
             xPosition2 += slope2;
             yPosition++;
+            this.colorInterpolator1.startColor.r += this.colorInterpolator1.colorSlope.r;
+            this.colorInterpolator1.startColor.g += this.colorInterpolator1.colorSlope.g;
+            this.colorInterpolator1.startColor.b += this.colorInterpolator1.colorSlope.b;
+
+            this.colorInterpolator3.startColor.r += this.colorInterpolator3.colorSlope.r;
+            this.colorInterpolator3.startColor.g += this.colorInterpolator3.colorSlope.g;
+            this.colorInterpolator3.startColor.b += this.colorInterpolator3.colorSlope.b;
 
             curz1 += zslope1;
             curz2 += zslope2;
