+++ src/rasterizer/TexturedTriangleRasterizer.ts
@@ -1,7 +1,7 @@
-import { Framebuffer } from '../../Framebuffer';
-import { Vector3f } from '../../math/Vector3f';
-import { Vector4f } from '../../math/Vector4f';
-import { Vertex } from '../../Vertex';
+import { Framebuffer } from '../Framebuffer';
+import { Vector3f } from '../math/Vector3f';
+import { Vector4f } from '../math/Vector4f';
+import { Vertex } from '../Vertex';
 
 export class TexturedTriangleRasterizer {
