--- src/rendering-pipelines/WireframeRenderingPipeline.ts
@@ -1,3 +1,0 @@
-export class WireframeRenderingPipeline {
-
-}
