+++ src/screen-space-clipping/LeftClipEdge.ts
@@ -20,7 +20,7 @@ export class LeftClipEdge extends AbstractClipEdge {
         const factor: number = (Framebuffer.minWindow.x - p1.projection.x) / (p2.projection.x - p1.projection.x);
         vertex.color = p2.color.sub(p1.color).mul(factor).add(p1.color);
         vertex.projection = new Vector4f(Framebuffer.minWindow.x,
-            p1.projection.y + (p2.projection.y - p1.projection.y) * factor,
+            Math.round(p1.projection.y + (p2.projection.y - p1.projection.y) * factor),
             1 / (1 / p1.projection.z + (1 / p2.projection.z - 1 / p1.projection.z) * factor));
         return vertex;
     }
